Removed upgrade integration tests
