+++ .github/workflows/ci.yaml
@@ -81,7 +81,6 @@ jobs:
           - password-rotation-integration
           - plugins-integration
           - tls-integration
-          - upgrade-integration
         agent-versions:
           - "2.9.44"  # renovate: latest juju 2
           - "3.1.5"  # renovate: latest juju 3
