+++ tests/integration/conftest.py
@@ -9,9 +9,7 @@ from pathlib import Path
 import boto3
 import pytest
 from pytest_operator.plugin import OpsTest
-from tenacity import Retrying, stop_after_delay, wait_fixed
 
-from tests.integration.ha_tests.conftest import APPLICATION_NAME
 from tests.integration.helpers import construct_endpoint
 
 AWS = "AWS"
@@ -60,22 +58,6 @@ async def cloud_configs(ops_test: OpsTest) -> None:
         # GCS doesn't support batch delete operation, so delete the objects one by one.
         for bucket_object in bucket.objects.filter(Prefix=config["path"].lstrip("/")):
             bucket_object.delete()
-
-
-@pytest.fixture()
-async def continuous_writes(ops_test: OpsTest) -> None:
-    """Deploy the charm that makes continuous writes to PostgreSQL."""
-    yield
-    # Clear the written data at the end.
-    for attempt in Retrying(stop=stop_after_delay(60 * 5), wait=wait_fixed(3), reraise=True):
-        with attempt:
-            action = (
-                await ops_test.model.applications[APPLICATION_NAME]
-                .units[0]
-                .run_action("clear-continuous-writes")
-            )
-            await action.wait()
-            assert action.results["result"] == "True", "Unable to clear up continuous_writes table"
 
 
 @pytest.fixture(scope="module")
