+++ tox.ini
@@ -174,17 +174,6 @@ commands =
     pip install juju=={env:LIBJUJU}
     poetry run pytest -v --tb native --log-cli-level=INFO -s  --durations=0 {posargs} {[vars]tests_path}/integration/test_tls.py
 
-[testenv:upgrade-integration-{juju2, juju3}]
-description = Run upgrade integration tests
-pass_env =
-    {[testenv]pass_env}
-    CI
-    CI_PACKED_CHARMS
-commands =
-    poetry install --with integration
-    pip install juju=={env:LIBJUJU}
-    poetry run pytest -v --tb native --log-cli-level=INFO -s  --durations=0 {posargs} {[vars]tests_path}/integration/test_upgrade.py
-
 [testenv:integration-{juju2, juju3}]
 description = Run all integration tests
 pass_env =
